+++ dusk-merkle/benches/blake3.rs
@@ -9,10 +9,39 @@ use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 use rand::rngs::StdRng;
 use rand::{RngCore, SeedableRng};
 
-use blake3::Hasher;
+use blake3::{Hash as Blake3Hash, Hasher};
 
-use dusk_merkle::blake3::Item;
-use dusk_merkle::Tree;
+use dusk_merkle::{Aggregate, Tree};
+
+const EMPTY_HASH: Item = Item([0; 32]);
+
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub struct Item([u8; 32]);
+
+impl From<Blake3Hash> for Item {
+    fn from(h: Blake3Hash) -> Self {
+        Self(h.into())
+    }
+}
+
+impl<const A: usize> Aggregate<A> for Item {
+    const EMPTY_SUBTREE: Self = EMPTY_HASH;
+
+    fn aggregate(items: [&Self; A]) -> Self {
+        let mut hasher = Hasher::new();
+        for item in items {
+            hasher.update(&item.0);
+        }
+        hasher.finalize().into()
+    }
+}
+
+impl Item {
+    #[must_use]
+    pub fn new(bytes: [u8; 32]) -> Self {
+        Item(bytes)
+    }
+}
 
 const H: usize = 32;
 const A: usize = 2;